Adding Comments
It still looks like you are not using eclipse to write your java code.  Make sure you use it so that you can know the compile errors. For example, on line 3 the curly brace is turned the wrong direction.  Eclipse will help you to know that. Dont forget to check your formatting as well.  If you correct this file and resubmit you can earn points back to this grade.  Keep up the good work.